Create first_draft.html
Probably need to scrap most of these .js, .html, .css files, but just forking this for now to see how this works. Maybe there's a better way to do this? Maybe I should be made a collaborator and then I could submit pull requests directly? Is that how it's done?